namespace veteran status model
