+++ app/models/emis_redis/veteran_status.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+require 'emis/veteran_status_service'
+
+module EMISRedis
+  class VeteranStatus < Model
+    CLASS_NAME = 'VeteranStatusService'
+    redis_config_key :veteran_status_response
+
+    def veteran?
+      raise VeteranStatus::Unauthorized unless @user.loa3?
+      any_veteran_indicator?(emis_response('get_veteran_status').items.first)
+    end
+
+    class NotAuthorized < StandardError
+    end
+
+    private
+
+    def any_veteran_indicator?(item)
+      item.post911_deployment_indicator == 'Y' ||
+        item.post911_combat_indicator == 'Y' ||
+        item.pre911_deployment_indicator == 'Y'
+    end
+  end
+end
+
