+++ spec/support/veteran_status/stub_veteran_status.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 def stub_veteran_status(is_veteran = true)
-  allow_any_instance_of(VeteranStatus).to receive(:veteran?).and_return(is_veteran)
+  allow_any_instance_of(EMISRedis::VeteranStatus).to receive(:veteran?).and_return(is_veteran)
 end
